lite/requirements.txt
Flask==3.0.3
gunicorn==23.0.0
PyYAML==6.0.2
requests==2.32.3
bcrypt==4.2.0
